code refactoring add modules and functions with arguments
